feat: Implement `CacheMissing` blockstore wrapper
