+++ car-mirror/src/push.rs
@@ -1,8 +1,8 @@
 use crate::{
+    cache::Cache,
     common::{block_receive, block_send, CarFile, Config, ReceiverState},
     error::Error,
     messages::PushResponse,
-    traits::Cache,
 };
 use libipld_core::cid::Cid;
 use wnfs_common::BlockStore;
@@ -51,13 +51,13 @@ pub async fn response(
 #[cfg(test)]
 mod tests {
     use crate::{
+        cache::NoCache,
         common::Config,
         dag_walk::DagWalk,
         test_utils::{
             get_cid_at_approx_path, setup_random_dag, total_dag_blocks, total_dag_bytes, Metrics,
             Rvg,
         },
-        traits::NoCache,
     };
     use anyhow::Result;
     use futures::TryStreamExt;
@@ -188,10 +188,10 @@ mod tests {
 #[cfg(test)]
 mod proptests {
     use crate::{
+        cache::NoCache,
         common::Config,
         dag_walk::DagWalk,
         test_utils::{setup_blockstore, variable_blocksize_dag},
-        traits::NoCache,
     };
     use futures::TryStreamExt;
     use libipld::{Cid, Ipld};
